fix: use constant when comparing positions
Let's set a constant so we can use 'is' comparison. Using '==' is not
sufficient since `False == 0` would return True, when we would want it to
be False if `0` is meant to be a position.

Although...there's a tiny likelihood (super close to 0%) that
someone would use this class (found in a repo for book exercise)
would decide to pass `False` as `position`. 🤔
It might be a good idea anyway to implement it this way to reinforce the idea
in my own head that `False == 0` returns `True`. 😅